+++ aesara/tensor/random/__init__.py
@@ -1,5 +1,5 @@
-# Initialize `RandomVariable` optimizations
-import aesara.tensor.random.opt
+# Initialize `RandomVariable` rewrites
+import aesara.tensor.random.rewriting
 import aesara.tensor.random.utils
 from aesara.tensor.random.basic import *
 from aesara.tensor.random.op import RandomState, default_rng
