+++ aesara/tensor/random/rewriting.py
